“Fix” and acknowledge failing tests
These specs, now marked as pending, were written wrong, and were testing
that the generated document was equal to *itself*, which is a pretty
blatant false positive.  Making them real tests revealed that the XML
is *not* equivalent.  I’m marking them as pending since I’m not sure
how major the differences are and whether or not we want a different
strategy for testing sufficient XML equivalency.